refactor: 주문 관리 및 메뉴 관리 컴포넌트 FSD 구조통합 및 불필요한 파일 삭제
