+++ src/app/order-manage/page.tsx
@@ -1,11 +1,13 @@
-import { OrderHistory } from "@/components/pages/order-manage/orderHistory/OrderHistory";
-import { OrderList } from "@/components/pages/order-manage/orderList/OrderList";
+import OrderManage from "@/views/OrderManage";
 
 export default function OrderManagePage() {
   return (
     <main className="flex h-screen w-full items-start p-24 bg-[#F2F2F0] gap-4">
-      <OrderList />
-      <OrderHistory />
+      {/* <div className="flex w-full h-full gap-4">
+        <OrderList />
+        <OrderHistory />
+      </div> */}
+      <OrderManage />
     </main>
   );
 }
