+++ src/app/page.tsx
@@ -1,9 +1,9 @@
-import MainPage from "@/views/MainPage";
+import MenuManage from "@/views/MainPage";
 
 export default function Home() {
   return (
     <main className="flex flex-col h-screen items-center justify-between p-24 bg-[#F2F2F0] ">
-      <MainPage />
+      <MenuManage />
     </main>
   );
 }
