--- src/components/pages/order-manage/orderList/EditOrder.tsx
@@ -1,248 +1,0 @@
-import { Button } from "@/components/ui/button";
-import {
-  Dialog,
-  DialogClose,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger,
-} from "@/components/ui/dialog";
-import { fetchWithAuth } from "@/lib/fetchWithAuth";
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-import { FormEvent, useState } from "react";
-
-import { MenuItemType } from "@/shared/types";
-import { OrderItemDatas, OrderListItems } from "../type/OrderItem";
-
-interface EditOrderProps {
-  orderItems: OrderItemDatas[];
-  orderId: string;
-}
-
-const editOrderState = async ({
-  orderId,
-  updateOrderItems,
-}: {
-  orderId: string;
-  updateOrderItems: OrderItemDatas[];
-}) => {
-  const items = updateOrderItems.map((item) => ({
-    menuId: +item.menu.id,
-    quantity: +item.quantity,
-    price: +item.menu.price,
-  }));
-
-  const response = await fetchWithAuth(`/api/order/${orderId}`, {
-    method: "PATCH",
-    body: JSON.stringify({
-      items,
-    }),
-    headers: {
-      "Content-Type": "application/json",
-    },
-  });
-  if (!response.ok) {
-    throw new Error("Failed to update orderItem state");
-  }
-  return response.json();
-};
-
-export const EditOrder = ({ orderItems, orderId }: EditOrderProps) => {
-  const [orderItemList, setOrderItemList] =
-    useState<OrderItemDatas[]>(orderItems);
-  const [showMenuList, setShowMenuList] = useState(false);
-  const queryClient = useQueryClient();
-
-  const {
-    isPending,
-    error,
-    data: menuList,
-    isLoading,
-  } = useQuery({
-    queryKey: ["menu-list"],
-    queryFn: async () => {
-      const fetchData = await fetchWithAuth("/api/menu");
-      const response: { data: MenuItemType[] } = await fetchData.json();
-      return response.data;
-    },
-  });
-  const mutation = useMutation({
-    mutationFn: editOrderState,
-    onMutate: async (updateItem) => {
-      await queryClient.cancelQueries({ queryKey: ["order-list"] });
-
-      const prevOrderItems = queryClient.getQueryData([
-        "order-list",
-      ]) as OrderListItems[];
-      const existIndex = prevOrderItems.findIndex(
-        (order) => order.id.toString() === updateItem.orderId
-      );
-      const updateOrderList = [...prevOrderItems];
-
-      updateOrderList[existIndex] = {
-        ...updateOrderList[existIndex],
-        items: updateItem.updateOrderItems,
-      };
-
-      queryClient.setQueryData(["order-list"], updateOrderList);
-
-      return { prevOrderItems };
-    },
-    onSettled: () => {
-      queryClient.invalidateQueries({ queryKey: ["order-list"] });
-    },
-    onError: (error, updateOrderId, context) => {
-      console.error("Failed to update:", error);
-      queryClient.setQueryData(["order-list"], context?.prevOrderItems);
-    },
-  });
-
-  const handleAddMenu = (menu: MenuItemType) => {
-    setOrderItemList((prev) => {
-      const exist = prev.find((item) => item.menu.id === menu.id.toString());
-      if (exist) {
-        // 이미 있으면 수량 +1
-        return prev.map((item) =>
-          item.menu.id === menu.id.toString()
-            ? { ...item, quantity: item.quantity + 1 }
-            : item
-        );
-      }
-      // 없으면 OrderItemDatas 타입에 맞게 추가
-      return [
-        ...prev,
-        {
-          id: Date.now(), // 임시 id
-          menu: {
-            ...menu,
-            id: menu.id.toString(),
-            quantity: 1,
-            created_at: new Date(),
-            updated_at: new Date(),
-          },
-          quantity: 1,
-          price: menu.price,
-        },
-      ];
-    });
-  };
-
-  const handleChangeItemQuantity = (id: number, quantity: number) => {
-    setOrderItemList((prev) =>
-      prev.map((item) => (item.id === id ? { ...item, quantity } : item))
-    );
-  };
-  const handleRemoveItem = (id: number) => {
-    setOrderItemList((prev) => prev.filter((item) => item.id !== id));
-  };
-
-  const submitHandler = async (e: FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    mutation.mutate({ orderId, updateOrderItems: orderItemList });
-  };
-
-  return (
-    <Dialog>
-      <DialogTrigger>수정</DialogTrigger>
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle>주문 수정</DialogTitle>
-          <DialogDescription></DialogDescription>
-        </DialogHeader>
-        <form onSubmit={submitHandler}>
-          <fieldset>
-            <legend className="font-bold mb-2">주문 목록</legend>
-            <div className="flex flex-col gap-3 mb-4">
-              {orderItemList.map((item) => (
-                <div
-                  key={item.id}
-                  className="flex items-center gap-4 bg-gray-50 rounded-lg px-4 py-2 shadow-sm"
-                >
-                  <span className="w-32 font-semibold">{item.menu.name}</span>
-                  <span className="w-20 text-right">{item.menu.price}원</span>
-                  <input
-                    className="border py-1 px-2 rounded-lg w-20 text-center"
-                    type="number"
-                    min={1}
-                    value={item.quantity}
-                    onChange={(e) =>
-                      handleChangeItemQuantity(item.id, Number(e.target.value))
-                    }
-                  />
-                  <button
-                    type="button"
-                    className="ml-auto text-gray-400 hover:text-red-500"
-                    onClick={() => handleRemoveItem(item.id)}
-                    aria-label="삭제"
-                  >
-                    X
-                  </button>
-                </div>
-              ))}
-            </div>
-            <div className="flex justify-between items-center mb-6">
-              <span className="text-lg font-bold text-brown-700">
-                총 가격:{" "}
-                {orderItemList
-                  .reduce((acc, cur) => acc + cur.menu.price * cur.quantity, 0)
-                  .toLocaleString()}
-                원
-              </span>
-              <Button
-                type="button"
-                variant="default"
-                className="ml-4"
-                onClick={() => setShowMenuList((prev) => !prev)}
-              >
-                메뉴추가
-              </Button>
-            </div>
-          </fieldset>
-
-          {showMenuList && menuList && menuList.length > 0 && (
-            <div className="border rounded-lg p-4 mb-6 bg-gray-50">
-              <div className="flex flex-wrap gap-4">
-                {menuList
-                  .filter((menu) => menu.is_active === 1)
-                  .map((menu) => (
-                    <div
-                      key={menu.id}
-                      className="flex flex-col items-center w-32 p-2 bg-white rounded shadow"
-                    >
-                      <div className="font-semibold mb-1">{menu.name}</div>
-                      <div className="text-sm text-gray-500 mb-2">
-                        {menu.price}원
-                      </div>
-                      <Button
-                        type="button"
-                        variant="outline"
-                        className="w-full"
-                        onClick={() => handleAddMenu(menu)}
-                      >
-                        추가
-                      </Button>
-                    </div>
-                  ))}
-              </div>
-            </div>
-          )}
-
-          <DialogFooter className="mt-8">
-            <DialogClose asChild>
-              <Button type="button" variant="secondary">
-                취소
-              </Button>
-            </DialogClose>
-            <DialogClose asChild>
-              <Button type="submit" variant="default">
-                변경하기
-              </Button>
-            </DialogClose>
-          </DialogFooter>
-        </form>
-      </DialogContent>
-    </Dialog>
-  );
-};
