--- src/components/pages/order-manage/orderList/OrderList.tsx
@@ -1,43 +1,0 @@
-"use client";
-import { fetchWithAuth } from "@/lib/fetchWithAuth";
-import { useOrderSSE } from "@/lib/useOrderSSE";
-import { useQuery } from "@tanstack/react-query";
-import { OrderListItems } from "../type/OrderItem";
-import { Order } from "./Order";
-
-export const OrderList = () => {
-  const { isPending, data, error } = useQuery({
-    queryKey: ["order-list"],
-    queryFn: async () => {
-      const fetchData = await fetchWithAuth("/api/order/today");
-      const response: { data: OrderListItems[] } = await fetchData.json();
-      return response.data;
-    },
-  });
-  useOrderSSE();
-
-  if (isPending) return <div>Loading...</div>;
-
-  return (
-    <div className="flex flex-col flex-grow-[5] basis-[50%] max-w-[50%] min-w-[50%] h-[calc(100vh-120px)] gap-4">
-      <div className="flex items-center justify-center relative px-2 py-4 bg-[#FDEACA] rounded-lg shrink-0">
-        <span className="text-2xl">주문 현황</span>
-      </div>
-      <ul className="grid grid-cols-3 gap-2 flex-1 overflow-y-auto">
-        {data &&
-          data.map((order) => {
-            return (
-              order.status === "IN_PROGRESS" && (
-                <li key={order.id} className="justify-self-center">
-                  <Order
-                    orderItems={order.items}
-                    orderId={order.id.toString()}
-                  />
-                </li>
-              )
-            );
-          })}
-      </ul>
-    </div>
-  );
-};
