+++ src/widgets/index.ts
@@ -1,1 +1,4 @@
-export { MenuList } from "./menuList/ui/MenuList";
+export { ConfirmedOrderList } from "./confirmed-order-list/ui/ConfirmedOrderList";
+export { InprogressOrderList } from "./inprogress-order-list/ui/InprogressOrderList";
+export { MenuCart } from "./menu-cart/ui/MenuCart";
+export { MenuList } from "./menu-list/ui/MenuList";
